Add security to WichtelEventController

+++ backend/src/main/java/org/example/backend/controller/WichtelEventController.java
@@ -4,9 +4,12 @@ package org.example.backend.controller;
 import lombok.RequiredArgsConstructor;
 import org.example.backend.model.WichtelEventDTO;
 import org.example.backend.model.WichtelParticipant;
+import org.example.backend.service.AuthService;
 import org.example.backend.service.WichtelEventService;
 import org.springframework.http.HttpStatus;
+import org.springframework.security.oauth2.client.authentication.OAuth2AuthenticationToken;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 import java.util.NoSuchElementException;
@@ -17,55 +20,88 @@ import java.util.NoSuchElementException;
 public class WichtelEventController {
 
     private final WichtelEventService service;
+    private final AuthService authService;
 
-    @PostMapping("/{organizerId}")
-    String createEvent(@PathVariable String organizerId){
+    @PostMapping
+    String createEvent(OAuth2AuthenticationToken authentication) {
+        String organizerId = authService.getUserFromAuthToken(authentication).getId();
         return service.createEmptyEvent(organizerId);
     }
 
     @GetMapping
-    List<WichtelEventDTO> findAll(){
+    List<WichtelEventDTO> findAll() {
         return service.findAll();
     }
 
     @GetMapping("/{id}")
-    WichtelEventDTO findById(@PathVariable String id){
+    WichtelEventDTO findById(@PathVariable String id) {
         return service.findById(id);
     }
 
     @PutMapping("/{id}")
-    WichtelEventDTO update(@RequestBody WichtelEventDTO dto, @PathVariable String id){
-        return service.update(dto,id);
+    WichtelEventDTO update(@RequestBody WichtelEventDTO dto, @PathVariable String id, OAuth2AuthenticationToken authentication) {
+        if (authService.isOrganizerOfEvent(authentication, id)) {
+            return service.update(dto, id);
+        } else {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
+        }
+
     }
 
     @DeleteMapping("/{id}")
-    void delete(@PathVariable String id){
-        service.deleteById(id);
+    void delete(@PathVariable String id, OAuth2AuthenticationToken authentication) {
+        if (authService.isOrganizerOfEvent(authentication, id)) {
+            service.deleteById(id);
+        } else {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
+        }
     }
 
     @PostMapping("/{eventId}/{participantId}")
-    WichtelEventDTO addParticipant(@PathVariable String eventId,@PathVariable String participantId){
-        return service.addParticipant(eventId,participantId);
+    WichtelEventDTO addParticipant(@PathVariable String eventId, @PathVariable String participantId, OAuth2AuthenticationToken authentication) {
+        if (authService.isOrganizerOfEvent(authentication, eventId) || authService.loggedInUserHasId(authentication, participantId)) {
+            return service.addParticipant(eventId, participantId);
+        } else {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
+        }
     }
 
     @PutMapping("/{eventId}/{participantId}")
-    WichtelEventDTO updateParticipant(@PathVariable String eventId, @RequestBody WichtelParticipant participant,@PathVariable String participantId){
-        return service.updateParticipant(eventId,participant,participantId);
+    WichtelEventDTO updateParticipant(@PathVariable String eventId, @RequestBody WichtelParticipant participant, @PathVariable String participantId, OAuth2AuthenticationToken authentication) {
+        if (authService.loggedInUserHasId(authentication, participantId)) {
+            return service.updateParticipant(eventId, participant, participantId);
+        } else {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
+        }
     }
 
     @DeleteMapping("/{eventId}/{participantId}")
-    WichtelEventDTO deleteParticipant(@PathVariable String eventId,@PathVariable String participantId){
-        return service.deleteParticipant(eventId,participantId);
+    WichtelEventDTO deleteParticipant(@PathVariable String eventId, @PathVariable String participantId, OAuth2AuthenticationToken authentication) {
+        if (authService.isOrganizerOfEvent(authentication, eventId) || authService.loggedInUserHasId(authentication, participantId)) {
+            return service.deleteParticipant(eventId, participantId);
+        } else {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
+        }
+
     }
 
     @PostMapping("/pairings/{eventId}")
-    WichtelEventDTO generatePairings(@PathVariable String eventId){
-        return service.generatePairings(eventId);
+    WichtelEventDTO generatePairings(@PathVariable String eventId, OAuth2AuthenticationToken authentication) {
+        if (authService.isOrganizerOfEvent(authentication, eventId)) {
+            return service.generatePairings(eventId);
+        } else {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
+        }
+
     }
 
     @GetMapping("/{eventId}/{participantId}")
-    WichtelParticipant getMyPairing(@PathVariable String eventId,@PathVariable String participantId){
-        return service.getPairingOfUser(eventId,participantId);
+    WichtelParticipant getMyPairing(@PathVariable String eventId, @PathVariable String participantId, OAuth2AuthenticationToken authentication) {
+        if (authService.loggedInUserHasId(authentication, participantId)) {
+            return service.getPairingOfUser(eventId, participantId);
+        } else {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
+        }
     }
 
     @ExceptionHandler(NoSuchElementException.class)
@@ -83,6 +119,12 @@ public class WichtelEventController {
     @ExceptionHandler(IllegalStateException.class)
     @ResponseStatus(HttpStatus.PRECONDITION_REQUIRED)
     public String handleIllegalStateException(IllegalStateException exception) {
+        return exception.getMessage();
+    }
+
+    @ExceptionHandler(IllegalCallerException.class)
+    @ResponseStatus(HttpStatus.UNAUTHORIZED)
+    public String handleIllegalCallerException(IllegalCallerException exception) {
         return exception.getMessage();
     }
 }
